requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
pyaes==1.6.1              # via telethon
pyasn1==0.4.7             # via rsa
pysocks==1.7.1
python-dotenv==0.10.3
rsa==4.0                  # via telethon
telethon==1.10.6
